// lib/poski-real/types.ts
export interface PoskiCredentials {
  clientId: string;
  passwordMd5: string;
  softwareKey: string;
}

export interface PoskiSession {
  sessionId: string;
  lastUpdated: Date;
}
// ...
export interface PoskiHashResponse {
  session_id: string;
}

export interface PoskiLoginResponse {
  success: boolean;
  user_id?: number;
  user_name?: string;
}